[Refactor] Split local model discovery out of task-models settings (#536)
* refactor: extract local provider model discovery from task-models

Move Ollama/vLLM/LiteLLM discovery, recommendation scoring, and
qualification into a dedicated module so settings commands stay thin.

* refactor: drop unused local discovery exports for knip

---------
